avoiding dangerous copy in lambda

+++ src/src/tree/src/TreeConcrete.cpp
@@ -40,7 +40,7 @@ namespace mt {
     }
 
     std::set<Node*> TreeConcrete::nearSet(const NodeState& state, const Nodes::const_reverse_iterator& delimiter) const {
-        float Tree_size = std::distance(delimiter, this->nodes.rend());
+		float Tree_size = static_cast<float>(std::distance(delimiter, this->nodes.rend()));
         float ray = this->problem.getGamma() * powf(logf(Tree_size) / Tree_size, 1.f / static_cast<float>( this->problem.getProblemSize() ));
         float dist_att;
         std::set<Node*> nearS;
